src/test/extension.test.ts
import * as assert from 'assert';
import * as fs from 'fs';
import * as path from 'path';
import * as vscode from 'vscode';

import { generateTestContent } from '../extension'; // Path to your extension

// For file system operations


suite('Swagger to API Test Conversion Tests', () => {

	// Path to your sample/test Swagger file
	const swaggerFilePath = path.join(__dirname, '..', 'test', 'fixtures', 'sampleSwagger.json'); // Example path, adjust


	test('Generates test script for valid Swagger file', async () => {
		// 1. Read the Swagger file content (replace with async file reading in VS Code extension)
		let swaggerData: any; // You'll define the type based on your Swagger structure
		try {
			const fileContent = fs.readFileSync(swaggerFilePath, 'utf8'); // Replace with VS Code async file read if necessary
			swaggerData = JSON.parse(fileContent);
		} catch (error) {
			assert.fail(`Error reading or parsing Swagger file: ${error}`);
			return;
		}



		// 2. Define expected output based on Swagger file.
		// Here, you create the *expected* test script string.
		const expectedTestContent = `
          // Expected test script content based on sampleSwagger.json 
          describe('GET /users', () => {
            it('should return a list of users', async () => {
              // ... your expected test logic ...
            });
          });
        `; // Customize this!



		// 3. Call your conversion function
		const methods = { get: swaggerData.paths['/users'].get }; // Extract GET method from your Swagger
		const generatedTestContent = await generateTestContent(
			'http://example.com', '/api', '/users', 'Users', methods, swaggerData
		);



		// 4. Assertions - check the generated script
		assert.ok(generatedTestContent.includes("describe('GET /users'"), 'Should contain describe block');
		assert.ok(generatedTestContent.includes("it('should return a list of users'"), 'Should contain it block');

		// Better, but more complex approach (using snapshot testing if your framework supports it)
		// assert.equal(generatedTestContent, expectedTestContent);  // Strict equality, best with snapshot testing


	});

	test('Handles invalid Swagger file gracefully', async () => {
		// Use a specially crafted "invalid" swagger.json to simulate errors.
		const invalidSwaggerPath = path.join(__dirname, '..', 'test', 'fixtures', 'invalidSwagger.json');

		let invalidSwaggerData: any;
		try {
			const fileContent = fs.readFileSync(invalidSwaggerPath, 'utf8'); // Replace if using VS Code file API
			invalidSwaggerData = JSON.parse(fileContent);
		} catch (error) {
			// If parsing fails due to the invalid JSON, you might already handle that as a success (depending on your implementation):
			return; // Or assert something specific about the error message if you're handling it differently.
		}


		// Call your conversion function with the invalid data.        
		const methods = { get: invalidSwaggerData.paths['/users'].get || {} }; // Handle potentially missing paths.
		try {
			await generateTestContent('http://example.com', '/api', '/users', 'Users', methods, invalidSwaggerData);
			assert.fail('Expected an error to be thrown for invalid Swagger.'); // If your function *should* throw an error.

		} catch (error) {
			// Check for a specific error or error type
			assert.ok(error instanceof Error); // Check for a generic error (adjust as needed).
			// Or be more specific: assert.strictEqual(error.message, 'Expected error message');
		}
	});




});

